+++ openmdao.main/src/openmdao/main/seqentialflow.py
@@ -1,7 +1,14 @@
 """ A workflow that allows the user to explicitly specify the execution
-order. """
+order. This workflow serves as the immediate base class for the two most
+important workflows: Dataflow and CyclicWorkflow."""
 
 from openmdao.main.workflow import Workflow
+
+try:
+    from numpy import ndarray, zeros
+except ImportError as err:
+    logging.warn("In %s: %r", __file__, err)
+    from openmdao.main.numpy_fallback import ndarray, zeros
 
 __all__ = ['SequentialWorkflow']
 
@@ -13,6 +20,11 @@ class SequentialWorkflow(Workflow):
         """ Create an empty flow. """
         self._names = []
         super(SequentialWorkflow, self).__init__(parent, scope, members)
+        
+        # Bookkeeping for calculating the residual.
+        self._severed_edges = None
+        self.res = None
+        self.bounds = None        
 
     def __iter__(self):
         """Returns an iterator over the components in the workflow."""
@@ -111,3 +123,157 @@ class SequentialWorkflow(Workflow):
     def clear(self):
         """Remove all components from this workflow."""
         self._names = []
+
+    def initialize_residual(self):
+        """Creates the array that stores the residual. Also returns the
+        number of edges.
+        """
+        edges = self.get_interior_edges()
+        
+        nEdge = 0
+        self.bounds = {}
+        for edge in edges:
+            
+            src = edge[0]
+            val = self.scope.get(src)
+            
+            # Floats get 1 element
+            if isinstance(val, float):
+                width = 1
+                
+            # Arrays are the product of their dimensions
+            elif isinstance(val, ndarray):
+                shape = val.shape
+                if len(shape) == 2:
+                    width = shape[0]*shape[1]
+                else:
+                    width = shape[0]
+                    
+            # Nothing else is supported.
+            else:
+                msg = "Variable %s is of type %s. " % (src, type(val)) + \
+                      "This type is not supported by the MDA Solver."
+                self.scope.raise_exception(msg, RuntimeError)
+                
+            self.bounds[edge] = (nEdge, nEdge+width)
+            nEdge += width
+            
+        # Initialize the residual vector on the first time through, and also
+        # if for some reason the number of edges has changed.
+        if self.res is None or nEdge != self.res.shape[0]:
+            self.res = zeros((nEdge, 1))
+            
+        return nEdge
+    
+    def calculate_residuals(self):
+        """Caclulate and return the vector of residuals based on the current
+        state of the system in our workflow."""
+        
+        for edge in self.get_interior_edges():
+            src, target = edge
+            src_val = self.scope.get(src)
+            target_val = self.scope.get(target)
+            
+            i1, i2 = self.bounds[edge]
+            self.res[i1:i2] = src_val - target_val
+    
+        return self.res
+    
+    def set_new_state(self, dv):
+        """Adds a vector of new values to the current model state at the
+        input edges.
+        
+        dv: ndarray (nEdge, 1)
+            Array of values to add to the model inputs.
+        """
+        
+        # Apply new state to model
+        for edge in self._severed_edges:
+            src, target = edge
+                
+            i1, i2 = self.bounds[edge]
+            old_val = self.scope.get(target)
+            
+            # Arrays
+            if i2-i1 > 1:
+                
+                # Arrays that are 2D column vectors
+                if old_val.shape[0] > old_val.shape[1]:
+                    new_val = old_val.T + dv[i1:i2]
+                    new_val = new_val.T
+                    
+                # 2D Row vectors
+                else:
+                    new_val = old_val + dv[i1:i2]
+                
+            # Floats
+            else:
+                new_val = old_val + float(dv[i1:i2])
+                
+            # Poke new value into the input end of the edge.
+            self.scope.set(target, new_val, force=True)
+            
+            # Prevent OpenMDAO from stomping on our poked input.
+            parts = target.split('.')
+            comp_name = parts[0]
+            var_name = '.'.join(parts[1:])
+            comp = self.scope.get(comp_name)
+            valids = comp._valid_dict
+            valids[var_name] = True
+            
+            #(An alternative way to prevent the stomping. This is more
+            #concise, but setting an output and allowing OpenMDAO to pull it
+            #felt hackish.)
+            #self.scope.set(src, new_val, force=True)
+
+    def matvecFWD(self, arg):
+        '''Callback function for performing the matrix vector product of the
+        workflow's full Jacobian with an incoming vector arg.'''
+        
+        # Bookkeeping dictionaries
+        inputs = {}
+        outputs = {}
+        
+        # Start with zero-valued dictionaries cotaining keys for all inputs
+        for comp in self.__iter__():
+            name = comp.name
+            inputs[name] = {}
+            outputs[name] = {}
+            
+        # Fill input dictionaries with values from input arg.
+        for edge in self.get_interior_edges():
+            src, target = edge
+            i1, i2 = self.bounds[edge]
+            
+            parts = src.split('.')
+            comp_name = parts[0]
+            var_name = '.'.join(parts[1:])
+            
+            outputs[comp_name][var_name] = arg[i1:i2]
+            inputs[comp_name][var_name] = arg[i1:i2]
+            
+            parts = target.split('.')
+            comp_name = parts[0]
+            var_name = '.'.join(parts[1:])
+            
+            inputs[comp_name][var_name] = arg[i1:i2]
+            
+        # Call ApplyJ on each component
+        for comp in self.__iter__():
+            name = comp.name
+            comp.applyJ(inputs[name], outputs[name])
+            
+        # Poke results into the return vector
+        result = zeros(len(arg))
+        for edge in self.get_interior_edges():
+            src, target = edge
+            i1, i2 = self.bounds[edge]
+        
+            parts = src.split('.')
+            comp_name = parts[0]
+            var_name = '.'.join(parts[1:])
+            
+            result[i1:i2] = outputs[comp_name][var_name]
+        
+        return result
+
